Overriding deprecated collider files
