+++ cmd/collider/dictionarydbcompare.go
